+++ TgaBuilderLib/Transitions/TransitionHelper.MixSeqmentedPictures.cs
@@ -1,5 +1,6 @@
 ﻿using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using static System.Net.WebRequestMethods;
 
 namespace TgaBuilderLib.Transitions
 {
@@ -14,14 +15,13 @@ namespace TgaBuilderLib.Transitions
             if (bgPixels.Length != tilePixels.Length)
                 throw new ArgumentException("Pixel arrays must have same length.");
 
-            var currentTileData = new List<TileSegment>(TileData.Count);
-            foreach (var tile in TileData)         
-                currentTileData.Add((TileSegment)tile.Clone());
+            var currentCentroids = new (float X, float Y)[Centroids.Length];
+            Array.Copy(Centroids, currentCentroids, Centroids.Length);
 
             var currentLabels = new int[Labels.Length];
             Array.Copy(Labels, currentLabels, Labels.Length);
 
-            if (currentLabels.Length > 0 && currentLabels.Max() > currentTileData.Count)
+            if (currentLabels.Length > 0 && currentLabels.Max() > currentCentroids.Length)
                 return bgPixels; // Fallback in case of race condition
 
             // Determine relevant edges based on mode and reverse pivot
@@ -29,7 +29,7 @@ namespace TgaBuilderLib.Transitions
                 GetDrawnEdgeTilesBools(Mode, ReversePivot);
 
             // Selection step: determine which pixels are drawn (Input → Label Map → Selection)
-            bool[] selection = BuildSelection(currentTileData, currentLabels,
+            bool[] selection = BuildSelection(currentCentroids, currentLabels,
                 checkTop, checkBottom, checkLeft, checkRight);
 
             byte[] result = new byte[bgPixels.Length];
@@ -90,47 +90,87 @@ namespace TgaBuilderLib.Transitions
         // The selection is the union of all qualified tiles' pixels, optionally filtered by
         // corner-slicing trigonometry as a pre-step when SliceCornerTiles is enabled.
         private bool[] BuildSelection(
-            List<TileSegment> tileData,
+            (float X, float Y)[] centroids,
             int[] labels,
             bool checkTop, bool checkBottom, bool checkLeft, bool checkRight)
         {
             bool[] selection = new bool[Width * Height];
+            int labelCount = centroids.Length;
 
-            // Optional pre-step: identify corner tiles for pixel-level trigonometric filtering.
-            // Passes drawn-edge flags so that each corner's dominant axis can be determined.
+            // --- STEP 1: Fast Grouping of Offsets (CSR Approach) ---
+            // 1. Count pixels per label
+            int[] counts = new int[labelCount + 1];
+            for (int i = 0; i < labels.Length; i++)
+            {
+                int lbl = labels[i];
+                if (lbl > 0 && lbl <= labelCount) counts[lbl]++;
+            }
+
+            // 2. Build Start Indices (Prefix Sum)
+            int[] startIndices = new int[labelCount + 2];
+            int currentPos = 0;
+            for (int i = 1; i <= labelCount; i++)
+            {
+                startIndices[i] = currentPos;
+                currentPos += counts[i];
+            }
+            startIndices[labelCount + 1] = currentPos;
+
+            // 3. Fill Flat Offsets array
+            // We use a temporary pointer array to track current write positions
+            int[] flatOffsets = new int[currentPos];
+            int[] writePos = (int[])startIndices.Clone();
+
+            for (int i = 0; i < labels.Length; i++)
+            {
+                int lbl = labels[i];
+                if (lbl > 0 && lbl <= labelCount)
+                {
+                    flatOffsets[writePos[lbl]++] = i;
+                }
+            }
+
+            // --- STEP 2: Selection Logic ---
             var cornerTileMap = SliceCornerTiles
                 ? BuildCornerTileMap(labels, checkTop, checkBottom, checkLeft, checkRight)
                 : null;
 
-            for (int i = 0; i < tileData.Count; i++)
+            for (int i = 0; i < labelCount; i++)
             {
-                var tile = tileData[i];
+                var centroid = centroids[i];
                 int labelID = i + 1;
 
+                // Get the slice of offsets for this specific label from our flat array
+                int start = startIndices[labelID];
+                int end = startIndices[labelID + 1];
+                ReadOnlySpan<int> tileOffsets = new ReadOnlySpan<int>(flatOffsets, start, end - start);
+
+                if (tileOffsets.Length == 0) continue;
+
                 // Pivot condition
-                float v = ComputeFocusV(Mode, tile);
+                float v = ComputeFocusV(Mode, centroid);
                 bool shouldDraw = ReversePivot ? (v <= Pivot) : (v >= Pivot);
 
                 // Avoid background-touching edge tiles being drawn
                 if (shouldDraw)
-                    shouldDraw = !DoesTileTouchRequiredEdge(tile, !checkTop, !checkBottom, !checkLeft, !checkRight);
+                    shouldDraw = !DoesTileTouchRequiredEdge(tileOffsets, !checkTop, !checkBottom, !checkLeft, !checkRight);
 
                 // Ensure required edge tiles are always drawn
                 if (!shouldDraw)
-                    shouldDraw = DoesTileTouchRequiredEdge(tile, checkTop, checkBottom, checkLeft, checkRight);
+                    shouldDraw = DoesTileTouchRequiredEdge(tileOffsets, checkTop, checkBottom, checkLeft, checkRight);
 
                 if (!shouldDraw)
                     continue;
 
-                // Corner slicing pre-step: for corner tiles, only add pixels that satisfy
-                // the trigonometric condition relative to the drawn edge axis.
+                // Corner slicing logic
                 if (cornerTileMap != null && cornerTileMap.TryGetValue(labelID, out var cornerInfo))
                 {
                     if (cornerInfo.drawsHoriz && cornerInfo.drawsVert)
                     {
-                        // Tile touches drawn edges on both axes — include all pixels (no slicing).
-                        foreach (int offset in tile.PixelOffsets)
+                        // Include all pixels in the tile
+                        foreach (int offset in tileOffsets)
                         {
+                            // Using original coordinate extraction logic
                             int py = offset / Stride;
                             int px = (offset % Stride) / TRANSITIONS_BPP;
                             selection[py * Width + px] = true;
@@ -139,14 +179,9 @@ namespace TgaBuilderLib.Transitions
                     else
                     {
                         float tanAngle = ComputeCornerSliceTanAngle(cornerInfo.cx, cornerInfo.cy);
-
-                        // Horizontal-dominant corner (top/bottom drawn edge): keep the region
-                        // closer to the horizontal edge — dy < dx * tanAngle.
-                        // Vertical-dominant corner (left/right drawn edge) or pivot-only:
-                        // keep the region closer to the vertical edge — dy >= dx * tanAngle.
                         bool keepHorizSide = cornerInfo.drawsHoriz;
 
-                        foreach (int offset in tile.PixelOffsets)
+                        foreach (int offset in tileOffsets)
                         {
                             int py = offset / Stride;
                             int px = (offset % Stride) / TRANSITIONS_BPP;
@@ -156,7 +191,7 @@ namespace TgaBuilderLib.Transitions
 
                             bool include = keepHorizSide
                                 ? (dy < dx * tanAngle)    // near horizontal (top/bottom) edge
-                                : (dy >= dx * tanAngle);  // near vertical (left/right) edge or default
+                                : (dy >= dx * tanAngle);  // near vertical (left/right) edge
 
                             if (include)
                                 selection[py * Width + px] = true;
@@ -165,7 +200,8 @@ namespace TgaBuilderLib.Transitions
                 }
                 else
                 {
-                    foreach (int offset in tile.PixelOffsets)
+                    // Standard tile (no corner slicing)
+                    foreach (int offset in tileOffsets)
                     {
                         int py = offset / Stride;
                         int px = (offset % Stride) / TRANSITIONS_BPP;
@@ -256,28 +292,28 @@ namespace TgaBuilderLib.Transitions
             bool reversePivot)
             => (mode, reversePivot) switch
             {
-                (TransitionMode.Top, false) => (true, false, false, false),
-                (TransitionMode.Top, true) => (false, true, true, true),
-                (TransitionMode.Bottom, false) => (false, true, false, false),
-                (TransitionMode.Bottom, true) => (true, false, true, true),
-                (TransitionMode.Left, false) => (false, false, true, false),
-                (TransitionMode.Left, true) => (true, true, false, true),
-                (TransitionMode.Right, false) => (false, false, false, true),
-                (TransitionMode.Right, true) => (true, true, true, false),
-                (TransitionMode.DiagonalTopLeft, false) => (true, false, true, false),
-                (TransitionMode.DiagonalTopLeft, true) => (false, true, false, true),
-                (TransitionMode.DiagonalTopRight, false) => (true, false, false, true),
-                (TransitionMode.DiagonalTopRight, true) => (false, true, true, false),
+                (TransitionMode.Top, false)                 => (true, false, false, false),
+                (TransitionMode.Top, true)                  => (false, true, true, true),
+                (TransitionMode.Bottom, false)              => (false, true, false, false),
+                (TransitionMode.Bottom, true)               => (true, false, true, true),
+                (TransitionMode.Left, false)                => (false, false, true, false),
+                (TransitionMode.Left, true)                 => (true, true, false, true),
+                (TransitionMode.Right, false)               => (false, false, false, true),
+                (TransitionMode.Right, true)                => (true, true, true, false),
+                (TransitionMode.DiagonalTopLeft, false)     => (true, false, true, false),
+                (TransitionMode.DiagonalTopLeft, true)      => (false, true, false, true),
+                (TransitionMode.DiagonalTopRight, false)    => (true, false, false, true),
+                (TransitionMode.DiagonalTopRight, true)     => (false, true, true, false),
                 _ => throw new ArgumentException("Invalid mode.")
             };
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         // Checks whether a tile touches any of the requested image edges.
-        private bool DoesTileTouchRequiredEdge(TileSegment tile, bool top, bool bottom, bool left, bool right)
+        private bool DoesTileTouchRequiredEdge(ReadOnlySpan<int> pixelOffsets, bool top, bool bottom, bool left, bool right)
         {
             int touchPixCount = 0;
 
-            foreach (int offset in tile.PixelOffsets)
+            foreach (int offset in pixelOffsets)
             {
                 int y = offset / Stride;
                 int x = (offset % Stride) / TRANSITIONS_BPP;
@@ -302,10 +338,10 @@ namespace TgaBuilderLib.Transitions
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         // Computes a normalized focus value for a tile based on its centroid.
-        private float ComputeFocusV(TransitionMode mode, TileSegment tile)
+        private float ComputeFocusV(TransitionMode mode, (float X, float Y) centroid)
         {
-            float nx = tile.CentroidX;
-            float ny = tile.CentroidY;
+            float nx = centroid.X;
+            float ny = centroid.Y;
 
             // --- Topological logic excerpt ---
             float distToT1 = 0, distToT2 = 0;
@@ -325,36 +361,12 @@ namespace TgaBuilderLib.Transitions
         private (float distToT1, float distToT2) ComputeTopologicy(TransitionMode mode, float nx, float ny)
             => mode switch
             {
-                TransitionMode.Top => (
-                    Math.Min(nx, Math.Min(1.0f - nx, 1.0f - ny)),
-                    ny
-                ),
-
-                TransitionMode.Bottom => (
-                    Math.Min(nx, Math.Min(1.0f - nx, ny)),
-                    1.0f - ny
-                ),
-
-                TransitionMode.Left => (
-                    Math.Min(ny, Math.Min(1.0f - ny, 1.0f - nx)),
-                    nx
-                ),
-
-                TransitionMode.Right => (
-                    Math.Min(ny, Math.Min(1.0f - ny, nx)),
-                    1.0f - nx
-                ),
-
-                TransitionMode.DiagonalTopLeft => (
-                    Math.Min(1.0f - nx, 1.0f - ny),
-                    Math.Min(nx, ny)
-                ),
-
-                TransitionMode.DiagonalTopRight => (
-                    Math.Min(nx, 1.0f - ny),
-                    Math.Min(1.0f - nx, ny)
-                ),
-
+                TransitionMode.Top              => (Math.Min(nx, Math.Min(1.0f - nx, 1.0f - ny)), ny),
+                TransitionMode.Bottom           => (Math.Min(nx, Math.Min(1.0f - nx, ny)), 1.0f - ny),
+                TransitionMode.Left             => (Math.Min(ny, Math.Min(1.0f - ny, 1.0f - nx)), nx),
+                TransitionMode.Right            => (Math.Min(ny, Math.Min(1.0f - ny, nx)), 1.0f - nx),
+                TransitionMode.DiagonalTopLeft  => (Math.Min(1.0f - nx, 1.0f - ny), Math.Min(nx, ny)),
+                TransitionMode.DiagonalTopRight => (Math.Min(nx, 1.0f - ny), Math.Min(1.0f - nx, ny)),
                 _ => (0f, 0f)
             };
     }
