+++ TgaBuilderLib/Transitions/TransitionHelper.cs
@@ -14,12 +14,13 @@ namespace TgaBuilderLib.Transitions
 
         public int[] Labels { get; private set; } = Array.Empty<int>();
 
+        public (float X, float Y)[] Centroids { get; set; } = Array.Empty<(float X, float Y)>();
+
         public int Width { get; set; }
         public int Height { get; set; }
         public int Stride { get; set; }
         public TransitionMode Mode { get; set; }
 
-        public List<TileSegment> TileData { get; set; } = new List<TileSegment>();
         public float Hardness { get; set; } = 0.5f;
         public float Pivot { get; set; } = 0.5f;
         public float Offset { get; set; } = 0f;
@@ -37,7 +38,7 @@ namespace TgaBuilderLib.Transitions
             LastAnalysisWidth = 0;
             LastAnalysisHeight = 0;
             Labels = Array.Empty<int>();
-            TileData.Clear();
+            Centroids = Array.Empty<(float X,float Y)>();
             Hardness = 0.5f;
             Pivot = 0.5f;
             Offset = 0f;
